Change the ESI data to be stored in the ESI url
When the ESI data was stored in the cache, it could be lost on a Magento cache
clear or when the cache data expired, with no way to regenerate it. By storing
it in the URL we get around this issue. A potential issue with this method is
the ESI URLs may end up being too long depending on how much data gets encoded
in the URL. Currently per-client URLs end up being around ~550 chars. If a URL
is more than 2047 characters a warning message will be logged.