Ensure POST requests get normalized data
Normalization only happens after potentially sending a request to the
backend. This generates hard to debug errors when descisions are made
based on these variables that affects applications behaviors.

For example, if one of the normalized values is used to select a
different storeview, then sending off the request prematurely causes the
wrong storeview to be selected.

Therefore we move all normalization to the earliest stage in vcl_recv
and it should remain before any `return(pipe)` statement is made, unless
explicitedly documented why it's safe to deviate from the straight and
narrow.