Add hash for persistent_shopping_cart cookie when hashing private blocks
This prevents the contents of private ESI blocks from being cached where there is no frontend cookie supplied with a request.

The persistent features allow the output of a customers name, or shopping cart contents within the header and it is possible for these to be seen by other users if this cookie is not included in the hash.