Fix issue with customer name not being removed from header when clearing persistent cookie
The default Magento functionality provides a "Not <name>?" link to allow a
customer to clear the persistent cookie on return to Magento.

Clicking this link with Turpentine enabled doesn't cause the header to be purged.

This fix adds the "persistent_session_expired" event to the default header's
flush_events. This event is dispatched by Mage_Persistent_IndexController::_cleanup()