Treat frontend_cid cookie like frontend with vcl_fix
There are various issues / stack exchange posts about "add to basket"
buttons not working the first time someone presses them when using the vcl_fix
method and running with an https insecure_base_url via say
nginx:443->varnish->nginx.

These issues are happening because magento is sending a second
cookie, frontend_cid, under this configuration. The value of
frontend_cid is kept in $_SESSION and the cookie is validated when the
session is started.

This patch adds the frontend_cid cookie to hash_data on esi private
access the same as frontend cookie and appears to work.

NOTES:
- I haven't worried about the case when not using vcl_fix
- X-Varnish-Set-Cookie shows frontend_cid for me, as beresp only sends one cookie