PHP7 Method Signature Inheritance Fix
PHP7 throws the following warning

> Warning: Declaration of Nexcessnet_Turpentine_Model_Varnish_Configurator_Version4::_vcl_director_backend($host, $port, $descriptor, $probeUrl = '', $options = Array) should be compatible with Nexcessnet_Turpentine_Model_Varnish_Configurator_Abstract::_vcl_director_backend($host, $port, $probeUrl = '', $options = Array)

Because Configurator_Version4 is redefining the method signature of
_vcl_director_backend. It makes no practical difference because it's
only called from _vcl_directory, which is also redefined, but PHP7 will
stop throwing the warning if the parameter has a default value. I also
moved it to the end to make it clear which parameter was new.