Add real_ip configuration option.
If:
* There is a proxy in front of varnish (e.g. https)
* The proxy is not connecting to varnish via the proxy protocol
(available in varnish >= 4.1)

client.ip will be the ip address of the proxy, which is probably not
correct, see: https://info.varnish-software.com/blog/failure-to-purge-a-story-about-client.ip-and-proxies

If the real_ip configuration option is enabled, then the vcl generated
will compare ACLs to std.ip(req.http.x-real-ip, "0.0.0.0").

X-Real-IP should be set by the upstream proxy.