Enable Adminhtml controller
This was introduced in https://github.com/nexcess/magento-turpentine/pull/817,
but this conflicted with the existing controller. This conflict was identified
in https://github.com/nexcess/magento-turpentine/issues/821 and resolved in
https://github.com/nexcess/magento-turpentine/pull/823 by removing this line.
Without this line, the overridden controller is not used (and therefore the
protection it adds is not being applied).

This change reintroduces the required configuration in a non-conflicting way.